feat(settings): give API keys their own nav section instead of a Profile sidebar box
Founder flagged from real dashboard screenshots that API key management
was cramped into a Profile sidebar card. Lift it into a new Settings.tsx
component with its own nav item (desktop sidebar + mobile bottom bar,
now 6 items), leaving Profile.tsx for identity/public-presence concerns.
