+++ app/components/Profile.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { CheckCircle2, Globe, KeyRound, Loader2, Save, ShieldCheck, Trash2, User, Copy, ExternalLink, Share2 } from 'lucide-react'
+import { CheckCircle2, Globe, Loader2, Save, ShieldCheck, User, Copy, ExternalLink, Share2 } from 'lucide-react'
 import { ReactNode, useEffect, useState } from 'react'
 import { useAuth } from '../contexts/AuthContext'
 import { authFetch, jsonAuthFetch } from '../lib/authFetch'
@@ -14,14 +14,6 @@ interface UserProfile {
   public_resource_count?: number
 }
 
-interface ApiKeySummary {
-  id: string
-  prefix: string
-  label: string | null
-  created_at: string
-  last_used_at: string | null
-}
-
 export function Profile() {
   const { user } = useAuth()
   const [profile, setProfile] = useState<UserProfile | null>(null)
@@ -32,12 +24,6 @@ export function Profile() {
   const [message, setMessage] = useState<{ type: 'success' | 'error'; text: string } | null>(null)
   const [copied, setCopied] = useState(false)
   const [shareUrl, setShareUrl] = useState('')
-
-  const [apiKeys, setApiKeys] = useState<ApiKeySummary[]>([])
-  const [loadingKeys, setLoadingKeys] = useState(false)
-  const [generatingKey, setGeneratingKey] = useState(false)
-  const [newRawKey, setNewRawKey] = useState<string | null>(null)
-  const [keyCopied, setKeyCopied] = useState(false)
 
   const handleCopyLink = async () => {
     if (!shareUrl) return
@@ -53,63 +39,8 @@ export function Profile() {
   useEffect(() => {
     if (user) {
       loadProfile()
-      loadApiKeys()
     }
   }, [user])
-
-  const loadApiKeys = async () => {
-    if (!user) return
-    setLoadingKeys(true)
-    try {
-      const response = await authFetch(user, '/api/settings/api-keys')
-      if (!response.ok) throw new Error('Failed to load API keys')
-      const data = await response.json()
-      setApiKeys(data.keys || [])
-    } catch (err) {
-      console.error('Error loading API keys:', err)
-    } finally {
-      setLoadingKeys(false)
-    }
-  }
-
-  const handleGenerateKey = async () => {
-    if (!user) return
-    setGeneratingKey(true)
-    setNewRawKey(null)
-    try {
-      const response = await jsonAuthFetch(user, '/api/settings/api-keys', { method: 'POST' })
-      const data = await response.json()
-      if (!response.ok) throw new Error(data.error || 'Failed to generate API key')
-      setNewRawKey(data.key)
-      loadApiKeys()
-    } catch (err) {
-      setMessage({ type: 'error', text: err instanceof Error ? err.message : 'Failed to generate API key' })
-    } finally {
-      setGeneratingKey(false)
-    }
-  }
-
-  const handleRevokeKey = async (id: string) => {
-    if (!user) return
-    try {
-      const response = await authFetch(user, `/api/settings/api-keys?id=${id}`, { method: 'DELETE' })
-      if (!response.ok) throw new Error('Failed to revoke API key')
-      setApiKeys((keys) => keys.filter((k) => k.id !== id))
-    } catch (err) {
-      setMessage({ type: 'error', text: err instanceof Error ? err.message : 'Failed to revoke API key' })
-    }
-  }
-
-  const handleCopyKey = async () => {
-    if (!newRawKey) return
-    try {
-      await navigator.clipboard.writeText(newRawKey)
-      setKeyCopied(true)
-      setTimeout(() => setKeyCopied(false), 2000)
-    } catch (err) {
-      console.error('Failed to copy API key:', err)
-    }
-  }
 
   useEffect(() => {
     if (profile?.username) {
@@ -305,62 +236,6 @@ export function Profile() {
               </div>
             </section>
           )}
-
-          <section className="app-panel p-4">
-            <div className="mb-3 flex items-center gap-2 text-sm font-bold text-zinc-950 dark:text-white">
-              <KeyRound className="h-4 w-4 text-blue-600" />
-              API Keys
-            </div>
-            <p className="mb-3 text-xs text-zinc-500 dark:text-zinc-400">
-              Use an API key for REST API access or the DumpIt MCP server. Keys are shown once — copy them somewhere safe.
-            </p>
-
-            {newRawKey && (
-              <div className="app-muted-panel mb-3 p-3">
-                <p className="mb-2 text-xs font-semibold text-emerald-700 dark:text-emerald-400">New key generated — copy it now:</p>
-                <div className="flex items-center gap-2">
-                  <code className="flex-1 truncate rounded bg-white px-2 py-1.5 text-xs dark:bg-zinc-900">{newRawKey}</code>
-                  <button
-                    onClick={handleCopyKey}
-                    className="inline-flex h-8 w-8 shrink-0 items-center justify-center rounded-lg border border-zinc-200 hover:bg-zinc-50 dark:border-zinc-800 dark:hover:bg-zinc-900"
-                  >
-                    <Copy className="h-3.5 w-3.5" />
-                  </button>
-                </div>
-                {keyCopied && <p className="mt-1 text-xs text-emerald-600">Copied!</p>}
-              </div>
-            )}
-
-            {loadingKeys ? (
-              <Loader2 className="h-4 w-4 animate-spin text-zinc-400" />
-            ) : apiKeys.length > 0 ? (
-              <ul className="mb-3 space-y-2">
-                {apiKeys.map((key) => (
-                  <li key={key.id} className="flex items-center justify-between rounded-lg border border-zinc-200 px-3 py-2 text-xs dark:border-zinc-800">
-                    <code className="text-zinc-600 dark:text-zinc-300">{key.prefix}</code>
-                    <button
-                      onClick={() => handleRevokeKey(key.id)}
-                      aria-label="Revoke key"
-                      className="text-zinc-400 hover:text-red-600"
-                    >
-                      <Trash2 className="h-3.5 w-3.5" />
-                    </button>
-                  </li>
-                ))}
-              </ul>
-            ) : (
-              <p className="mb-3 text-xs text-zinc-400 dark:text-zinc-500">No API keys yet.</p>
-            )}
-
-            <button
-              onClick={handleGenerateKey}
-              disabled={generatingKey || apiKeys.length >= 5}
-              className="inline-flex min-h-9 w-full items-center justify-center gap-1.5 rounded-lg border border-zinc-200 hover:bg-zinc-50 dark:border-zinc-800 dark:hover:bg-zinc-900 text-xs font-bold text-zinc-700 dark:text-zinc-200 transition-colors disabled:cursor-not-allowed disabled:opacity-50"
-            >
-              {generatingKey ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <KeyRound className="h-3.5 w-3.5" />}
-              Generate new key
-            </button>
-          </section>
 
           {profile && (
             <section className="app-panel p-4">
